Long polling implementation.
This commit is the final implementation for integrating the long polling  mechanism for the message search label.  However its still needs to be tested in real world scenerio where request will transfer to multiple nodes. Till then I am checking this as the final implementation. Also note that this commit uses a scala diode rertry mechanism.