Completed the message post and retrieval
This commit is another major commit to complete the message posting and the retrieval. Prior to this message were being posted in wrong format resulting in exceptions during retrieval and parsing.
Also this commit intend to solve a bug related to not able to get the message for both to and from the logged in user. Only thing that remains at this point is the date parsing for the created and modified fields and also sending special character with the json.