Added small doc for the search label logic.
This commit is for documenting what is really happening in the searches
label logic. Confusing part is the updateLabel method of the searches
handler which really deals with the toggle the state of the label tree on
searches view. Same file contains logic for creating the logic families
which is documented as well.