Changing message model.
This is a major commit which includes a lot of architectural changes
primarily for the messages post and retrieve. The model binding to the
message post view and message retrieve view is changed heavily to make the
architectural more generic so as to make the content post and retieval
more of a plug and play architecture.