Removing session pinger
Earlier for the session ping relay we relied on the session pinger architecture which relied on a subscription on a particular flag on session root model. This is removed as of now and now it is reverted to good old recursive call in the refresh messages action.

Subsquent tests are all also modified accordingly.