Moving latest features from synereo
This commit includes a number of changes. First of all it removes the complicated pinger logic and replaces it with usual recurssive session ping. This  should reduce the complexity and confusion and unnecessary ping due to errornous response. Also this include the error handling for the srp implementation which handles a number of error responses.