Refactor in content utils
This commit includes the refactoring of the process response method of the content utils. Now it is modified to return seq of response content instead of post making it open for use in future responses which may derive from response content. Also it includes the modifications in the sort content so as to handle the responses for the balance changed which earlier was handled with error prone string match. The unit tests are modified accordingly.