Fixing modal popups on main menu.
Earlier when user made a transition between different url the modal popups remained opened as the main menu is not updated. Now the logic for opening and closing of these popups is moved into app module and app root modal and the popups are explicitly closed by dispatching an action once user navigates to different location. Also the show and hide of the right section is now handled more gracefully.