Implementation for server details post
This commit includes the initial implementation for the server details on the node server menu. It includes boopickle and autowire libraries because they simplify a lot of stuff related to serialisation and deserialisation and the octet stream http transfer. Probably it won't be used in the main branch as in the main branch we use normal ajax call without cross server-client dtos.