Implementation for I18N
This commit includes the working implementation for I18N. Current implementation includes an ajax call (getLang) from server which then reads the json files stored on the server. Once this json is received on client it is then parsed into js.Dynamic with JSON.parse. For each user we need to include a property in state say lang whose initial state is read from Synereo circuit I18N root model. Each view should have a method which subscribe to the changes in the root model and updates the state property accordingly. Individual elements should read the property from the state like state.lang.getDynamic("watch_video").asInstanceOf[String]

Scope of improvement: This is working technique however there are few issues here.

1. You need to include a lang property in state for each view since each individual view  listens to its state. Initial state should get value from the root model creating many duplicates for the same object
2. You need to subscribe to changes in the rootmodel for each view and then update the model accordingly increasing the boilerplate code.

Possible resolution:
There can be a generic solution where you can basically transfer the change in the root model to the child elements. In that case we can subscribe to root model in the app module and transfer the changes to the child elements. Will investigate on this one.