Fixing no show for no cnxn json
This commit fixes the no show for the messages with no connection in the
strigified json. The version before dashboard optimisation were based on
wrong assumptions. For no connections in the json it was earlier assumed
that sender is "me" and receiver is "UNKNOWN" which is wrong. Essentially
since we have no info of the sender and receiver in the json both are of
them are "UNKNOWN". By default server will never send the messages with no
connections because ui has a restriction where you can't messages to empty
connections. So we will always get a connection in the json. However since
it is required to support the json with no connection which can be present
in the imported json we have this workaround now in place. There are few
issues related to these unknown sender and receiver right now:

1. Sender and receiver will show up as UNKNOWN.
2. Reply to post and forward post will not work.
3. Amplification of post will not work since we do not have the session uri for the sender.

Further steps to be taken:
Move the connections from the stringified json to the evalSubscribeResponse json in the connection field which currently holds only single connection